requirements.txt
selenium
beautifulsoup4
pandas
Pillow
requests